requirements.txt
asgiref==3.8.1
Django==4.2.1
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==11.0.0
sqlparse==0.5.3
tzdata==2024.2
drf-yasg==1.21.7